Consolidate private intra-subsystem capabilities.
